requirements.txt
Flask
mysql-connector-python
flask-cors
pylint
pytest
pandas
pyjwt
pydantic[email]
bcrypt
SQLAlchemy
flask-marshmallow
Flask-Mail